+++ monkey/common/event_queue/__init__.py
@@ -1,3 +1,3 @@
 from .types import EventSubscriber
 from .i_agent_event_queue import IAgentEventQueue
-from .pypubsub_event_queue import PyPubSubAgentEventQueue
+from .pypubsub_agent_event_queue import PyPubSubAgentEventQueue
